backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.0
python-dotenv>=1.0
Pillow>=10.0
